Crate bolero_generator

Re-exports

pub use either;

pub use driver::Driver;

pub use one_of::one_of;

pub use one_of::one_value_of;

Modules

alloc_generators

array

atomic

bool

bounded

char

combinator

driver

num

one_of

prelude

range

result

std_generators

time

tuple

Macros

impl_key_values_collection_generator

impl_values_collection_generator

Structs

Constant

TypeValueGenerator

Non-parameterized ValueGenerator given a TypeGenerator

Traits

TypeGenerator

Generate a value for a given type

TypeGeneratorWithParams

Convert a type generator into the default value generator

ValueGenerator

Generate a value with a parameterized generator

Functions

constant

Always generate the same value

gen

Generate a value for a given type

gen_with

Generate a value for a given type with additional constraints

Derive Macros

TypeGenerator

Derive the an implementation of TypeGenerator for the given type.